案例3-设备维保图文混答知识库/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
ragflow_sdk==0.20.4
PyMuPDF==1.23.8
python-dotenv==1.0.0
Pillow==11.1.0
boto3==1.34.0